extends symlink-creator to work with new symlinks (#25)
provides a whitelist of directory names to the `symlink_creator.sh`. symlinks will only be created for `.yml` files contained within the whitelisted directories.

the extended implementation is too complicated and somewhat hacky and thus should be improved with another solution that preferably does not use a bash script. the issue for this required improvement is created here: https://github.com/graknlabs/web-dev/issues/24