Delete component selector and move search bar into the freed-up space (#183)
## What is the goal of this PR?

Previously the component selector was displayed above the navigation.
Now the component selector is removed and search is moved into that
space.

## What are the changes implemented in this PR?

As all the components currently are displayed in the navigation and
there is only one version of components. It's safe to remove component
selector. This also makes additional space for the search which wasn't
looking good full width above the content and ToC.