Prevent adding trailing slash to direct parent url (#198)
## What is the goal of this PR?

Due to defect of default UI relativize path - when creating one level
parent path it was always resolved to directory (with trailing slash).
Now it is resolved correctly to path without trailing slash if
destination is a file, and with trailing slash if destination is a
directory. (docs don't use directories in the current configuration)

## What are the changes implemented in this PR?

When resolving relative path underscore is appended to always produce
meaningful path. Underscore is removed after resolving relative path.
From `/a/b` to `/a` it resolves to:
`/a` => `/a_` => `../a_` => `../a`
instead of old
`/a` => ` ` => `.`
From `/a/b` to `/a/` it resolves to:
`/a/` => `/a/_` => `_` => ` ` => `.` => `./`
which has the same result as old
`/a/` => ` ` => `.` => `./`