Update typedb-driver referenced branches after their renaming (#235)
## What is the goal of this PR?

Fix TypeDB Driver API Reference pages after they got broken due to the
renaming of the `typedb-driver` repo's branches.

## What are the changes implemented in this PR?

Update referenced branches to `[master, "2.29.2"]`. `master` is the
`3.x` version, and `"2.29.2"` is the `2.x` version now.
Note that 2.29.2 is [a
tag](https://github.com/typedb/typedb-driver/tree/2.29.2), but it is
expected to work just like a branch (hopefully). For cleanness purposes,
we don't push a separate `2.x` branch, but it can be done if needed.