Fix 2.x driver API ref rendering (#237)
## Release notes: product changes
Use the `2.x` branch referencing instead of the `2.29.2` tag for the 2.x
TypeDB Driver API Reference page.

## Motivation
Based on the current state of the website, it does not use the tagging
correctly, trying to apply 3.x documentation to the page. This fix will
allow us to maintain the 2.x docs, adding more updates and bugfixes.

## Implementation
Add `2.x` to `branches` instead of specifying `2.29.2` in `tags`. The
branch is [here](https://github.com/typedb/typedb-driver/tree/2.x).